convert setHTMLSync to normal function and move it to index.js
